refactor(profile): Replace local DTOs with shared service versions
Contributes to profile/geo functionality

- Removed local DTOs from Geo service
  * CountryDto
  * RegionDto
  * DistrictDto
  * CityDto
  * StreetDto
  * HouseDto
- Using DTOs from Shared service instead
- Updated dependent classes
