+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/city/service/CityService.java
@@ -1,7 +1,7 @@
 package ru.pavbatol.myplace.geo.city.service;
 
-import ru.pavbatol.myplace.geo.city.dto.CityDto;
 import ru.pavbatol.myplace.shared.dto.pagination.SimpleSlice;
+import ru.pavbatol.myplace.shared.dto.profile.geo.city.CityDto;
 
 public interface CityService {
     CityDto create(CityDto dto);
