+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/country/controller/AdminCountryController.java
@@ -9,8 +9,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.pavbatol.myplace.app.util.Marker;
-import ru.pavbatol.myplace.geo.country.dto.CountryDto;
 import ru.pavbatol.myplace.geo.country.service.CountryService;
+import ru.pavbatol.myplace.shared.dto.profile.geo.country.CountryDto;
 
 import javax.validation.Valid;
