+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/country/service/CountryService.java
@@ -1,7 +1,7 @@
 package ru.pavbatol.myplace.geo.country.service;
 
-import ru.pavbatol.myplace.geo.country.dto.CountryDto;
 import ru.pavbatol.myplace.shared.dto.pagination.SimpleSlice;
+import ru.pavbatol.myplace.shared.dto.profile.geo.country.CountryDto;
 
 public interface CountryService {
     CountryDto create(CountryDto dto);
