--- mp-profile/src/main/java/ru/pavbatol/myplace/geo/district/dto/DistrictDto.java
@@ -1,23 +1,0 @@
-package ru.pavbatol.myplace.geo.district.dto;
-
-import lombok.Value;
-import ru.pavbatol.myplace.app.util.Marker;
-import ru.pavbatol.myplace.geo.common.IdentifiableGeo;
-import ru.pavbatol.myplace.geo.region.dto.RegionDto;
-
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Null;
-import javax.validation.constraints.Size;
-
-@Value
-public class DistrictDto implements IdentifiableGeo {
-    @Null(groups = Marker.OnCreate.class)
-    Long id;
-
-    @NotNull(groups = Marker.OnCreate.class)
-    RegionDto region;
-
-    @NotNull(groups = Marker.OnCreate.class)
-    @Size(max = 255)
-    String name;
-}
