+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/district/mapper/DistrictMapper.java
@@ -2,12 +2,12 @@ package ru.pavbatol.myplace.geo.district.mapper;
 
 import org.mapstruct.*;
 import ru.pavbatol.myplace.app.util.Checker;
-import ru.pavbatol.myplace.geo.district.dto.DistrictDto;
 import ru.pavbatol.myplace.geo.district.model.District;
-import ru.pavbatol.myplace.geo.region.dto.RegionDto;
 import ru.pavbatol.myplace.geo.region.mapper.RegionMapper;
 import ru.pavbatol.myplace.geo.region.model.Region;
 import ru.pavbatol.myplace.geo.region.repository.RegionRepository;
+import ru.pavbatol.myplace.shared.dto.profile.geo.district.DistrictDto;
+import ru.pavbatol.myplace.shared.dto.profile.geo.region.RegionDto;
 
 @Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE, componentModel = "spring", uses = RegionMapper.class)
 public interface DistrictMapper {
