--- mp-profile/src/main/java/ru/pavbatol/myplace/geo/house/dto/HouseDto.java
@@ -1,30 +1,0 @@
-package ru.pavbatol.myplace.geo.house.dto;
-
-import lombok.Value;
-import ru.pavbatol.myplace.app.util.Marker;
-import ru.pavbatol.myplace.geo.common.IdentifiableGeo;
-import ru.pavbatol.myplace.geo.street.dto.StreetDto;
-
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Null;
-import javax.validation.constraints.PositiveOrZero;
-import javax.validation.constraints.Size;
-
-@Value
-public class HouseDto implements IdentifiableGeo {
-    @Null(groups = Marker.OnCreate.class)
-    Long id;
-
-    @NotNull(groups = Marker.OnCreate.class)
-    StreetDto street;
-
-    @NotNull(groups = Marker.OnCreate.class)
-    @Size(max = 10)
-    String number;
-
-    @PositiveOrZero
-    Double lat;
-
-    @PositiveOrZero
-    Double lon;
-}
