+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/house/mapper/HouseMapper.java
@@ -2,12 +2,12 @@ package ru.pavbatol.myplace.geo.house.mapper;
 
 import org.mapstruct.*;
 import ru.pavbatol.myplace.app.util.Checker;
-import ru.pavbatol.myplace.geo.house.dto.HouseDto;
 import ru.pavbatol.myplace.geo.house.model.House;
-import ru.pavbatol.myplace.geo.street.dto.StreetDto;
 import ru.pavbatol.myplace.geo.street.mapper.StreetMapper;
 import ru.pavbatol.myplace.geo.street.model.Street;
 import ru.pavbatol.myplace.geo.street.repository.StreetRepository;
+import ru.pavbatol.myplace.shared.dto.profile.geo.house.HouseDto;
+import ru.pavbatol.myplace.shared.dto.profile.geo.street.StreetDto;
 
 @Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE, componentModel = "spring", uses = StreetMapper.class)
 public interface HouseMapper {
