+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/region/controller/PrivateRegionController.java
@@ -7,9 +7,9 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import ru.pavbatol.myplace.geo.region.dto.RegionDto;
 import ru.pavbatol.myplace.geo.region.service.RegionService;
 import ru.pavbatol.myplace.shared.dto.pagination.SimpleSlice;
+import ru.pavbatol.myplace.shared.dto.profile.geo.region.RegionDto;
 
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
