--- mp-profile/src/main/java/ru/pavbatol/myplace/geo/region/dto/RegionDto.java
@@ -1,23 +1,0 @@
-package ru.pavbatol.myplace.geo.region.dto;
-
-import lombok.Value;
-import ru.pavbatol.myplace.app.util.Marker;
-import ru.pavbatol.myplace.geo.common.IdentifiableGeo;
-import ru.pavbatol.myplace.geo.country.dto.CountryDto;
-
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Null;
-import javax.validation.constraints.Size;
-
-@Value
-public class RegionDto implements IdentifiableGeo {
-    @Null(groups = Marker.OnCreate.class)
-    Long id;
-
-    @NotNull(groups = Marker.OnCreate.class)
-    CountryDto country;
-
-    @NotNull(groups = Marker.OnCreate.class)
-    @Size(max = 255)
-    String name;
-}
