+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/region/mapper/RegionMapper.java
@@ -2,12 +2,12 @@ package ru.pavbatol.myplace.geo.region.mapper;
 
 import org.mapstruct.*;
 import ru.pavbatol.myplace.app.util.Checker;
-import ru.pavbatol.myplace.geo.country.dto.CountryDto;
 import ru.pavbatol.myplace.geo.country.mapper.CountryMapper;
 import ru.pavbatol.myplace.geo.country.model.Country;
 import ru.pavbatol.myplace.geo.country.repository.CountryRepository;
-import ru.pavbatol.myplace.geo.region.dto.RegionDto;
 import ru.pavbatol.myplace.geo.region.model.Region;
+import ru.pavbatol.myplace.shared.dto.profile.geo.country.CountryDto;
+import ru.pavbatol.myplace.shared.dto.profile.geo.region.RegionDto;
 
 @Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE, componentModel = "spring", uses = CountryMapper.class)
 public interface RegionMapper {
