+++ mp-profile/src/main/java/ru/pavbatol/myplace/geo/street/service/StreetService.java
@@ -1,7 +1,7 @@
 package ru.pavbatol.myplace.geo.street.service;
 
-import ru.pavbatol.myplace.geo.street.dto.StreetDto;
 import ru.pavbatol.myplace.shared.dto.pagination.SimpleSlice;
+import ru.pavbatol.myplace.shared.dto.profile.geo.street.StreetDto;
 
 public interface StreetService {
     StreetDto create(StreetDto dto);
