+++ mp-profile/src/main/java/ru/pavbatol/myplace/profile/dto/ProfileDtoUpdate.java
@@ -2,8 +2,8 @@ package ru.pavbatol.myplace.profile.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Value;
-import ru.pavbatol.myplace.geo.house.dto.HouseDto;
 import ru.pavbatol.myplace.profile.model.Gender;
+import ru.pavbatol.myplace.shared.dto.profile.geo.house.HouseDto;
 
 import javax.validation.constraints.AssertTrue;
 import javax.validation.constraints.Email;
